Documentation
My biggest project in computer programming was creating my own crossword puzzle. It was a project assigned to me in CS 230 and it dealt with a lot of different programming codes in Python. To start me off I typed different steps to guide me through the project as you can see on each line with a hashtag. From line 10 to 24 the programming code I typed was for python to determine how many box spaces I needed for each word in the crossword puzzle. From line 26 to 36 the code I typed was to create the borders for the x and y columns so it actually created the boxes in crossword puzzle. Basically this coding helped make the design of a crossword puzzle. From line 40 to 52 I created my if-else statement so that way whoever was to play the game, each time they entered a word the program would determine if it's the correct letter or the wrong one. From lines 55 to 67 the code was to finish off the crossword puzzle with a border at the bottom that way whoever participated in the game and solved all the words would be notified that they completed the game.